packages/type-r-t2i/pyproject.toml
[project]
name = "type-r-t2i"
version = "0.1.0"
description = "Type-R text-to-image package."
readme = "README.md"
authors = [
    { name = "Wataru Shimoda", email = "wataru_shimoda@cyberagent.co.jp" },
]
requires-python = ">=3.10, <3.13"
dependencies = [
    "diffusers>=0.32.2",
    "loguru>=0.7.3",
    "openai>=1.65.5",
    "pillow>=10.4.0",
    "requests>=2.32.3",
    "torch>=2.1.2",
    "accelerate>=0.34.2",
    "google-genai>=1.19.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["type_r"]  # src layout does not work in editable mode.
